+++ services/PasswdService/include/IoT/Passwd/UserSerializer.h
@@ -20,8 +20,6 @@
 #include "IoT/Passwd/GroupDeserializer.h"
 #include "IoT/Passwd/GroupSerializer.h"
 #include "IoT/Passwd/PasswdService.h"
-#include "IoT/Passwd/UserDeserializer.h"
-#include "IoT/Passwd/UserSerializer.h"
 #include "Poco/RemotingNG/TypeSerializer.h"
 
 
@@ -45,9 +43,9 @@ public:
 		remoting__staticInitBegin(REMOTING__NAMES);
 		static const std::string REMOTING__NAMES[] = {"gid","groups","id","name","nic","passwordHash",""};
 		remoting__staticInitEnd(REMOTING__NAMES);
-		TypeSerializer<size_t >::serialize(REMOTING__NAMES[0], value.gid, ser);
+		TypeSerializer<unsigned long >::serialize(REMOTING__NAMES[0], value.gid, ser);
 		TypeSerializer<std::vector < IoT::PasswdService::Group > >::serialize(REMOTING__NAMES[1], value.groups, ser);
-		TypeSerializer<size_t >::serialize(REMOTING__NAMES[2], value.id, ser);
+		TypeSerializer<unsigned long >::serialize(REMOTING__NAMES[2], value.id, ser);
 		TypeSerializer<std::string >::serialize(REMOTING__NAMES[3], value.name, ser);
 		TypeSerializer<std::string >::serialize(REMOTING__NAMES[4], value.nic, ser);
 		TypeSerializer<std::string >::serialize(REMOTING__NAMES[5], value.passwordHash, ser);
